+++ rockcraft.yaml
@@ -1,4 +1,4 @@
-name: webhook-gateway
+name: mayfly
 base: bare
 build-base: ubuntu@24.04
 version: '0.1'
