+++ webhook-gateway/tox.ini
@@ -22,8 +22,8 @@ pass_env =
 
 
 
-[testenv:integration]
-description = Run integration tests
+[testenv:webhook-gateway-integration]
+description = Run webhook gateway integration tests
 deps =
     pytest
     pytest-operator
